# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Folio – Personal Notes API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: folders
    description: Folder management
  - name: notes
    description: Note management
  - name: noteVersions
    description: Note version history
  - name: stats
    description: App statistics

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/user:
    get:
      operationId: getCurrentAuthUser
      tags: [auth]
      summary: Get current authenticated user
      parameters:
        - $ref: "#/components/parameters/AuthorizationSessionHeader"
        - $ref: "#/components/parameters/SessionCookie"
      responses:
        "200":
          description: Auth status resolved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserEnvelope"

  /login:
    get:
      operationId: beginBrowserLogin
      tags: [auth]
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        "302":
          description: Redirect to OIDC authorization endpoint

  /callback:
    get:
      operationId: handleBrowserLoginCallback
      tags: [auth]
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
      responses:
        "302":
          description: Redirect to returnTo path on success

  /logout:
    get:
      operationId: logoutBrowserSession
      tags: [auth]
      summary: Clear session and begin OIDC logout
      parameters:
        - $ref: "#/components/parameters/AuthorizationSessionHeader"
        - $ref: "#/components/parameters/SessionCookie"
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
            default: /
      responses:
        "302":
          description: Redirect to OIDC provider logout URL

  /mobile-auth/token-exchange:
    post:
      operationId: exchangeMobileAuthorizationCode
      tags: [auth]
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MobileTokenExchangeRequest"
      responses:
        "200":
          description: Session created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MobileTokenExchangeSuccess"
        "400":
          description: Missing or invalid required fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Token did not contain usable claims
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Token exchange failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /mobile-auth/logout:
    post:
      operationId: logoutMobileSession
      tags: [auth]
      summary: Delete a mobile session token
      parameters:
        - $ref: "#/components/parameters/AuthorizationSessionHeader"
      responses:
        "200":
          description: Logout processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogoutSuccess"

  /folders:
    get:
      operationId: listFolders
      tags: [folders]
      summary: List all folders for the current user
      responses:
        "200":
          description: List of folders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Folder"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createFolder
      tags: [folders]
      summary: Create a new folder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FolderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Folder"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /folders/{id}:
    patch:
      operationId: updateFolder
      tags: [folders]
      summary: Update a folder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FolderPatch"
      responses:
        "200":
          description: Updated folder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Folder"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteFolder
      tags: [folders]
      summary: Delete a folder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List notes, optionally filtered by folder
      parameters:
        - name: folderId
          in: query
          required: false
          schema:
            type: integer
        - name: pinned
          in: query
          required: false
          schema:
            type: boolean
        - name: deleted
          in: query
          required: false
          schema:
            type: boolean
        - name: sortBy
          in: query
          required: false
          schema:
            type: string
            enum: [manual, updatedAt]
      responses:
        "200":
          description: List of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createNote
      tags: [notes]
      summary: Create a new note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /notes/reorder:
    post:
      operationId: reorderNotes
      tags: [notes]
      summary: Update the manual sort order of notes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderInput"
      responses:
        "200":
          description: Updated notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"

  /notes/{id}:
    get:
      operationId: getNote
      tags: [notes]
      summary: Get a single note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateNote
      tags: [notes]
      summary: Update a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotePatch"
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Soft-delete a note (move to trash)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Soft-deleted note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /notes/{id}/restore:
    post:
      operationId: restoreNote
      tags: [notes]
      summary: Restore a soft-deleted note from trash
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Restored note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /notes/{id}/purge:
    delete:
      operationId: purgeNote
      tags: [notes]
      summary: Permanently delete a note from trash
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Permanently deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /notes/{id}/pin:
    post:
      operationId: pinNote
      tags: [notes]
      summary: Pin a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pinned note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    delete:
      operationId: unpinNote
      tags: [notes]
      summary: Unpin a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unpinned note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"

  /notes/{id}/move:
    post:
      operationId: moveNote
      tags: [notes]
      summary: Move a note to a different folder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoveNoteInput"
      responses:
        "200":
          description: Moved note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /notes/{id}/versions:
    get:
      operationId: listNoteVersions
      tags: [noteVersions]
      summary: List version history for a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of note versions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NoteVersion"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /notes/{id}/versions/{versionId}/restore-copy:
    post:
      operationId: restoreNoteVersionAsCopy
      tags: [noteVersions]
      summary: Restore a note version as a new copy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: versionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: New note copy created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /stats:
    get:
      operationId: getStats
      tags: [stats]
      summary: Get app-wide statistics for the current user
      responses:
        "200":
          description: Statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: "Opaque session token — Bearer <sid>."
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string

  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    AuthUser:
      type: object
      required: [id]
      properties:
        id:
          type: string
        email:
          type: ["string", "null"]
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        profileImageUrl:
          type: ["string", "null"]

    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: "#/components/schemas/AuthUser"
            - type: "null"

    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1

    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string

    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean

    Folder:
      type: object
      required: [id, name, userId, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        userId:
          type: string
        parentFolderId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    FolderInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        parentFolderId:
          type: ["integer", "null"]

    FolderPatch:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        parentFolderId:
          type: ["integer", "null"]

    Note:
      type: object
      required: [id, title, content, userId, isPinned, isDeleted, sortOrder, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        userId:
          type: string
        folderId:
          type: ["integer", "null"]
        isPinned:
          type: boolean
        isDeleted:
          type: boolean
        deletedAt:
          type: ["string", "null"]
          format: date-time
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    NoteInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
        folderId:
          type: ["integer", "null"]
        sortOrder:
          type: integer

    NotePatch:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
        folderId:
          type: ["integer", "null"]
        isPinned:
          type: boolean
        sortOrder:
          type: integer

    ReorderInput:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/NoteOrderItem"

    NoteOrderItem:
      type: object
      required: [id, sortOrder]
      properties:
        id:
          type: integer
        sortOrder:
          type: integer

    MoveNoteInput:
      type: object
      required: [folderId]
      properties:
        folderId:
          type: ["integer", "null"]

    NoteVersion:
      type: object
      required: [id, noteId, content, createdAt]
      properties:
        id:
          type: integer
        noteId:
          type: integer
        content:
          type: string
        createdAt:
          type: string
          format: date-time

    AppStats:
      type: object
      required: [totalNotes, totalFolders, pinnedNotes, trashedNotes, recentlyEdited]
      properties:
        totalNotes:
          type: integer
        totalFolders:
          type: integer
        pinnedNotes:
          type: integer
        trashedNotes:
          type: integer
        recentlyEdited:
          type: array
          items:
            $ref: "#/components/schemas/Note"