# lib/api-zod/src/generated/types/appStats.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Folio – Personal Notes API
 * OpenAPI spec version: 0.1.0
 */
import type { Note } from './note';

export interface AppStats {
  totalNotes: number;
  totalFolders: number;
  pinnedNotes: number;
  trashedNotes: number;
  recentlyEdited: Note[];
}